disable LTO on x86_64 and znver1
LTO in combination with Clang results in segfault when trying to use AVX2 asm which is going to get picked by default on every znver1 machine and also significant amount of x86_64 machines (2013 and newer). I didn't manage to reproduce this on 32-bit.
Other option to consider would be to switch to GCC (where LTO doesnt break) but with -O3 which would give significant boost over Clang (gcc -O2 is slower than clang -O2).